requirements.txt
# Dependencies for real Supabase integration
supabase==2.0.0
python-dotenv==1.0.0
